backend/test/requirements.txt
pytest~=8.1.1
requests~=2.31.0
azure-cosmos~=4.6.0
azure-storage-queue~=12.10.0
